+++ website/react-magma-docs/src/components/CodeBlock/context.ts
@@ -4,6 +4,7 @@ import { Language } from 'prism-react-renderer';
 interface CodeBlockContextValue {
   language: Language;
   noRender: boolean;
+  noBorder: boolean;
   noCode: boolean;
   noCopy: boolean;
   noCodeSandbox: boolean;
@@ -14,6 +15,7 @@ interface CodeBlockContextValue {
 export const CodeBlockContext = React.createContext<CodeBlockContextValue>({
   language: 'typescript',
   title: 'Code Example',
+  noBorder: false,
   noCode: false,
   noCopy: false,
   noCodeSandbox: false,