+++ website/react-magma-docs/CHANGELOG.md
@@ -3,6 +3,78 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+### 2.1.8-alpha.9 (2021-03-15)
+
+**Note:** Version bump only for package react-magma-docs
+
+
+
+
+
+### 2.1.8-alpha.8 (2021-03-15)
+
+**Note:** Version bump only for package react-magma-docs
+
+
+
+
+
+### 2.1.8-alpha.7 (2021-03-15)
+
+**Note:** Version bump only for package react-magma-docs
+
+
+
+
+
+### 2.1.8-alpha.6 (2021-03-12)
+
+**Note:** Version bump only for package react-magma-docs
+
+
+
+
+
+### 2.1.8-alpha.5 (2021-03-12)
+
+**Note:** Version bump only for package react-magma-docs
+
+
+
+
+
+### 2.1.8-alpha.4 (2021-03-11)
+
+**Note:** Version bump only for package react-magma-docs
+
+
+
+
+
+### 2.1.8-alpha.3 (2021-03-11)
+
+**Note:** Version bump only for package react-magma-docs
+
+
+
+
+
+### 2.1.8-alpha.2 (2021-03-09)
+
+**Note:** Version bump only for package react-magma-docs
+
+
+
+
+
+### 2.1.8-alpha.1 (2021-03-09)
+
+**Note:** Version bump only for package react-magma-docs
+
+
+
+
+
 ### 2.1.8-alpha.0 (2021-03-05)
 
 **Note:** Version bump only for package react-magma-docs